refactor: extract shared agent stream core to eliminate duplication
Extract the ~800 lines of shared logic between chat-handler.ts and
agent-step.ts into a new agent-stream-core.ts module. Both files are
now thin wrappers over createAgentStreamExecute(), reducing agent-step
from 967 to 133 lines and chat-handler from 1306 to 389 lines.

Also fixes several gaps in the workflow path vs the serverless path:
- Adopt UsageTracker (was using raw variables, missing cache tokens,
  sandbox cost tracking, and usage record logging)
- Add missing tools/providerOptions args to runSummarizationStep
- Track summarization usage (cache read/write tokens, cost)
- Pass onToolCost callback to createTools for per-tool cost tracking
